Reset GeometryType.
